change token type

+++ mpi/ring.c
@@ -5,7 +5,7 @@
 
 
 int main(int argc, char *argv[]) {
-  int64_t token;
+  unsigned long long token;
   MPI_Init(&argc, &argv);
   int worldSize;
   MPI_Comm_size(MPI_COMM_WORLD, &worldSize);
@@ -14,7 +14,7 @@ int main(int argc, char *argv[]) {
   if (worldRank != 0) {
     MPI_Recv(&token, 1, MPI_INT, worldRank - 1, 0, MPI_COMM_WORLD,
              MPI_STATUS_IGNORE);
-    printf("process %d received token %ld from process %d\n", worldRank, token,
+    printf("process %d received token %lli from process %d\n", worldRank, token,
            worldRank - 1);
   } else {
     token = 1;
@@ -29,12 +29,10 @@ int main(int argc, char *argv[]) {
              MPI_COMM_WORLD);
   }
 
-  
-
   if (worldRank == 0) {
     MPI_Recv(&token, 1, MPI_INT, worldSize - 1, 0, MPI_COMM_WORLD,
              MPI_STATUS_IGNORE);
-    printf("process %d received token %ld from process %d\n", worldRank,
+    printf("process %d received token %lli from process %d\n", worldRank,
            token, worldSize - 1);
   }
